+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix error from parse_text_block on some case.
 - Fix filter bug on :text node.
 - Fix parse_interpolate_line bug.
+- Fix haml comment parser bug.
 
 ## 0.3.0 - 2022-01-26